requirements.txt
openai
pinecone-client
ragas
datasets
tqdm
PyMuPDF
python-docx
streamlit
langchain
